requirements.txt
pygame-2.1.2
gym